options and fontface actions
