+++ src/containers/Surah/index.js
@@ -44,7 +44,7 @@ import { surahsConnect, ayahsConnect } from './connect';
 
 import * as AudioActions from '../../redux/actions/AudioPlayerActions.js';
 import * as AyahActions from '../../redux/actions/AyahsActions.js';
-import * as OptionsActions from '../../redux/modules/options';
+import * as OptionsActions from '../../redux/actions/OptionsActions.js';
 
 const style = require('./style.scss');
