+++ src/redux/constants/FontFaceActionTypes.js
@@ -1,0 +1,1 @@
+export const LOAD = '@@quran/fontFaces/LOAD';
