+++ src/redux/constants/OptionsActionTypes.js
@@ -1,0 +1,2 @@
+export const TOGGLE_READING_MODE = '@@quran/options/TOGGLE_READING_MODE';
+export const SET_OPTION = '@@quran/options/SET_OPTION';
